+++ src/Modules/Language/Http/LanguageApiHandler.php
@@ -17,6 +17,7 @@ declare(strict_types=1);
 
 namespace Lwt\Modules\Language\Http;
 
+use Lwt\Modules\Language\Domain\ParserSelection;
 use Lwt\Modules\Language\Domain\WordSpacing;
 use Lwt\Shared\Infrastructure\Database\QueryBuilder;
 use Lwt\Shared\Infrastructure\Database\Settings;
@@ -76,7 +77,7 @@ class LanguageApiHandler implements ApiRoutableInterface
     public function getReadingConfiguration(int $langId): array
     {
         $record = QueryBuilder::table('languages')
-            ->select(['LgName', 'LgTTSVoiceAPI', 'LgRegexpWordCharacters', 'LgPiperVoiceId'])
+            ->select(['LgName', 'LgTTSVoiceAPI', 'LgRegexpWordCharacters', 'LgPiperVoiceId', 'LgParserType'])
             ->where('LgID', '=', $langId)
             ->firstPrepared();
 
@@ -103,7 +104,7 @@ class LanguageApiHandler implements ApiRoutableInterface
             $readingMode = "piper";
         } elseif ($ttsVoiceApi !== '') {
             $readingMode = "external";
-        } elseif (WordSpacing::usesMecabMagicWord($regexpWordChars)) {
+        } elseif (ParserSelection::rowTokenizesWithMecab($record)) {
             $readingMode = "internal";
         } else {
             $readingMode = "direct";